requirements.txt
numpy
scipy
matplotlib
opencv-python
pyserial
pyqt6
pyqt6-tools